Added mixin just for epsilon separation
