+++ unified_planning/model/mixins/__init__.py
@@ -14,6 +14,7 @@
 #
 
 from unified_planning.model.mixins.actions_set import ActionsSetMixin
+from unified_planning.model.mixins.epsilon_separation import EpsilonSeparationMixin
 from unified_planning.model.mixins.fluents_set import FluentsSetMixin
 from unified_planning.model.mixins.objects_set import ObjectsSetMixin
 from unified_planning.model.mixins.user_types_set import UserTypesSetMixin
@@ -23,6 +24,7 @@ from unified_planning.model.mixins.metrics import MetricsMixin
 
 __all__ = [
     "ActionsSetMixin",
+    "EpsilonSeparationMixin",
     "FluentsSetMixin",
     "ObjectsSetMixin",
     "UserTypesSetMixin",
