+++ unified_planning/model/scheduling/scheduling_problem.py
@@ -1,8 +1,13 @@
 from collections import OrderedDict
+from fractions import Fraction
 from typing import Optional, List, Union, Dict
 
 from unified_planning.model.expression import ConstantExpression
-from unified_planning.model.mixins import InitialStateMixin, MetricsMixin
+from unified_planning.model.mixins import (
+    InitialStateMixin,
+    MetricsMixin,
+    EpsilonSeparationMixin,
+)
 from unified_planning.model.mixins.objects_set import ObjectsSetMixin
 from unified_planning.model.mixins.fluents_set import FluentsSetMixin
 from unified_planning.model.mixins.user_types_set import UserTypesSetMixin
@@ -24,6 +29,7 @@ from unified_planning.model.timing import GlobalStartTiming
 class SchedulingProblem(  # type: ignore[misc]
     AbstractProblem,
     UserTypesSetMixin,
+    EpsilonSeparationMixin,
     FluentsSetMixin,
     ObjectsSetMixin,
     InitialStateMixin,
@@ -51,6 +57,7 @@ class SchedulingProblem(  # type: ignore[misc]
     ):
         AbstractProblem.__init__(self, name, environment)
         UserTypesSetMixin.__init__(self, self.environment, self.has_name)
+        EpsilonSeparationMixin.__init__(self, default=Fraction(1))
         FluentsSetMixin.__init__(
             self, self.environment, self._add_user_type, self.has_name, initial_defaults
         )
@@ -150,6 +157,7 @@ class SchedulingProblem(  # type: ignore[misc]
         UserTypesSetMixin._clone_to(self, new_p)
         ObjectsSetMixin._clone_to(self, new_p)
         FluentsSetMixin._clone_to(self, new_p)
+        EpsilonSeparationMixin._clone_to(self, new_p)
         InitialStateMixin._clone_to(self, new_p)
         MetricsMixin._clone_to(self, new_p, new_actions=None)
